more fields added. but the LogParams json always changes
